Fix behaviour on invalid URLs
